pyproject.toml
[project]
name = "madr"
version = "0.1.0"
description = ""
authors = [{ name = "Joao-Pedro-P-Holanda", email = "joaopedroph.dev@gmail.com" }]
requires-python = ">=3.11"
readme = "README.md"
dependencies = [
    "fastapi[standard] (>=0.121.2,<0.122.0)",
    "sqlalchemy[asyncio] (>=2.0.41,<3.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "pwdlib[argon2] (>=0.2.1,<0.3.0)",
    "alembic (>=1.16.2,<2.0.0)",
    "pydantic-extra-types (>=2.10.5,<3.0.0)",
    "psycopg-binary (>=3.2.9,<4.0.0)",
    "psycopg (>=3.2.9,<4.0.0)",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "taskipy>=1.14.1,<2",
    "ruff>=0.12.1,<0.13",
    "pytest>=8.4.1,<9",
    "aiosqlite>=0.21.0,<0.22",
    "freezegun>=1.5.2,<2",
    "factory-boy>=3.3.3,<4",
    "pytest-asyncio>=1.1.0,<2",
    "coverage>=7.10.4,<8",
]

[tool.uv]
package = false

[tool.fastapi]
entrypoint= "madr.app:app"

[tool.taskipy.tasks]
test = "pytest"
lint = "ruff check --fix ."
